Use new version of layers control minimap with included leaflet.sync, show first map after filtering if it's a baselayer
