+++ preview/index.html
@@ -6,7 +6,7 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
 
-	<link rel="stylesheet" href="http://cdn.leafletjs.com/leaflet-0.7/leaflet.css" />
+	<link rel="stylesheet" href="http://cdn.leafletjs.com/leaflet-0.7.3/leaflet.css" />
 
 	<style>
 		html {
@@ -77,10 +77,9 @@
 	</div>
 	<div id="map" class="map"></div>
 
-	<script src="http://cdn.leafletjs.com/leaflet-0.7/leaflet-src.js"></script>
+	<script src="http://cdn.leafletjs.com/leaflet-0.7.3/leaflet-src.js"></script>
 	<script src="../leaflet-providers.js"></script>
 
-	<script src="L.Map.Sync.js"></script>
 	<script src="L.Control.Layers.Minimap.js"></script>
 
 	<script src="highlightjs/highlight.pack.js"></script>
